test(get_histfile_path): Fix tests and improve error handling
The tests now run all together, when they are ran using only one thread.
This can accomplished by running them using
`cargo test -- -test-threads=1`.

Also improve error handling by not panicking every time a variable
could not be read. It is entirely plausible that one or more environment
variables are not present and therefore cannot be 'restored' after.

So if an environment variable cannot be read because it does not exist,
it is simply ignored.